Restructure project directories
