+++ frontend/src/components/editor/ScreenplayManager.tsx
