+++ frontend/src/components/editor/ToolsSidebar.tsx
