+++ frontend/src/components/editor/UserBar.tsx
