+++ frontend/src/components/editor/character-panel.tsx
