+++ frontend/src/components/editor/editor-sidebar.tsx
