+++ frontend/src/components/editor/editor-toolbar.tsx
