+++ frontend/src/components/editor/error-checker.tsx
