+++ frontend/src/components/editor/ruler.tsx
