+++ frontend/src/components/editor/script-navigator.tsx
