+++ frontend/src/components/editor/stash-panel.tsx
