+++ frontend/src/components/shared/Auth.tsx
