+++ frontend/src/components/shared/ErrorBoundary.tsx
