+++ frontend/src/pages/AiAgentsPage.tsx
