+++ frontend/src/pages/TextAnalysisPage.tsx
